+++ src/KingmakerGunslinger/Classes/NimbleArmor.cs
@@ -1,0 +1,11 @@
+namespace KingmakerGunslinger.Classes
+{
+    internal enum NimbleArmor
+    {
+        Unknown = 0,
+        None = 1,
+        Light = 2,
+        Medium = 3,
+        Heavy = 4
+    }
+}